Added fields for packet_id and transaction_id

+++ src/lorawan/helper/transactional-sender-helper.h
@@ -45,9 +45,9 @@ public:
 
   uint8_t GetPartialSignaturePacketSize (void) const;
 
-  void SetPacketsPerTransaction (uint32_t packets);
+  void SetPacketsPerTransaction (uint16_t packets);
 
-  uint32_t GetPacketsPerTransaction (void) const;
+  uint16_t GetPacketsPerTransaction (void) const;
 
   /**
    * Set the intra-transaction-delay
@@ -90,7 +90,7 @@ private:
   /**
    * The number of packets per transaction
    */
-  uint32_t packetsPerTransaction;
+  uint16_t packetsPerTransaction;
 
   /**
    * The interval between two consecutive transactions
